MM-17: Updates to flow types
