--- src/shared/reducers/types.js
@@ -1,4 +1,0 @@
-// @flow
-
-// @TODO: properly type the redux store
-export type State = Object;
