+++ src/shared/types/deviceinfo.types.js
@@ -1,16 +1,21 @@
 // @flow
-type Device = 'android' | 'ios' | 'chrome' | 'firefox' | 'desktop';
+import type { UserOptions } from './option.types';
+
+type DeviceOS = 'android' | 'ios' | 'chrome' | 'firefox' | 'desktop';
 
 export type DeviceInfo = {
   miner_id?: string,
-  cidr?: string,
-  asn?: string,
   lat?: string,
   lon?: string,
-  device_type: Device
+  device_type: DeviceOS
 };
 
 export type DeviceOptions = {
   wallet: string,
   wifi_only: boolean
+};
+
+export type Device = {
+  info: DeviceInfo,
+  options: UserOptions
 };
